Homework 6 -Main Package refactor -toString more methods.
