+++ src/Main.java
@@ -1,18 +1,18 @@
-import TechnicType.Brand.Apple;
-import TechnicType.Brand.Protected_Smartphone;
-import TechnicType.Brand.Samsung;
-import TechnicType.Brand.Xiaomi;
-import TechnicType.Builder.HuaweiBuilder;
+import com.mobile.store.Brand.Apple;
+import com.mobile.store.Brand.Protected_Smartphone;
+import com.mobile.store.Brand.Samsung;
+import com.mobile.store.Brand.Xiaomi;
+import com.mobile.store.Builder.HuaweiBuilder;
 
 public class Main {
 
 
     public static void main(String[] args) {
-        var iPhone = new Apple("TechnicType.Brand.Apple", "15 Pro Max", "iOs", 2023, 5.77, 512, "ApplePay", false, 1, 0);
-        var Samsung = new Samsung("TechnicType.Brand.Samsung", "S23 Ultra", "Android", 2023, 6.8, 512, "GooglePay", false, 2, 1024);
-        var SamsungProtected = new Protected_Smartphone("TechnicType.Brand.Samsung", "XCover 6 Pro", "Android", 2022, 6.6, 128, "GooglePay", true, 2, 1024, "MIL-STD-810H", true);
-        var Huawei = new HuaweiBuilder().setBrand("TechnicType.Brand.Huawei").setModel("Mate X3").setOperationSystem("HarmonyOS").setYearReleased(2023).setScreenSize(7.6).setBuildInMemoryStorage(512).setPaymentSystem("GooglePay").setAudioJackPresence(true).setSimCardNumSupport(2).setMaxSdCardSizeSupportGb(2024).createHuawei();
-        var Xiaomi = new Xiaomi("TechnicType.Brand.Xiaomi", "13 Ultra", "Android", 2023, 6.73, 512, "GooglePay", true, 2, 2024);
+        var iPhone = new Apple("Apple", "15 Pro Max", "iOs", 2023, 5.77, 512, "ApplePay", false, 1, 0);
+        var Samsung = new Samsung("Samsung", "S23 Ultra", "Android", 2023, 6.8, 512, "GooglePay", false, 2, 1024);
+        var SamsungProtected = new Protected_Smartphone("Samsung", "XCover 6 Pro", "Android", 2022, 6.6, 128, "GooglePay", true, 2, 1024, "MIL-STD-810H", true);
+        var Huawei = new HuaweiBuilder().setBrand("Huawei").setModel("Mate X3").setOperationSystem("HarmonyOS").setYearReleased(2023).setScreenSize(7.6).setBuildInMemoryStorage(512).setPaymentSystem("GooglePay").setAudioJackPresence(true).setSimCardNumSupport(2).setMaxSdCardSizeSupportGb(2024).createHuawei();
+        var Xiaomi = new Xiaomi("Xiaomi", "13 Ultra", "Android", 2023, 6.73, 512, "GooglePay", true, 2, 2024);
         var HuaweiBuilder = new HuaweiBuilder();
 
         System.out.println("Smartphones list:");
